+++ lizard_reportgenerator/views.py
@@ -1,22 +1,19 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
-from django.conf import settings
 from lizard_area.models import Area
 from lizard_reportgenerator.models import ReportTemplate
 from lizard_reportgenerator.models import GeneratedReport
 from lizard_security.models import DataSet
 
 from django.shortcuts import render_to_response
-from django.http import Http404
 from django.template import RequestContext
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
+from django.utils import encoding
 
 import xlwt as excel
 import urllib2
-from StringIO import StringIO
 import csv
 
-from cgi import escape
 import cStringIO as StringIO
 import ho.pisa as pisa
 
@@ -24,8 +21,6 @@ import ho.pisa as pisa
 def index(request, template='lizard_reportgenerator/index.html'):
     '''
         returns overview page of reports and historic pages
-
-
     '''
 
     #todo: make function less KRW specific
@@ -37,7 +32,8 @@ def index(request, template='lizard_reportgenerator/index.html'):
         data_sets = DataSet.objects.all()
     else:
         data_set_ids = getattr(request, 'ALLOWED_DATA_SET_IDS', [None])
-        data_sets = DataSet.objects.filter(permission_mappers__user_group__members=request.user).distinct()
+        data_sets = DataSet.objects.filter(
+            permission_mappers__user_group__members=request.user).distinct()
 
     aan_afvoergebied = None
     krw_gebied = None
@@ -67,7 +63,8 @@ def index(request, template='lizard_reportgenerator/index.html'):
         'selection_id': 'select_krw_waterlichaam',
         'area': krw_gebied,
         'report_templates': ReportTemplate.objects.filter(kind='krw'),
-        'generated_reports': GeneratedReport.objects.filter(template__kind='krw', area = krw_gebied)
+        'generated_reports': GeneratedReport.objects.filter(template__kind='krw',
+                                                            area=krw_gebied)
     })
 
     report_bloks.append({
@@ -76,11 +73,11 @@ def index(request, template='lizard_reportgenerator/index.html'):
         'selection_id': 'select_aan_afvoergebied',
         'area': aan_afvoergebied,
         'report_templates': ReportTemplate.objects.filter(kind='aan_afvoer'),
-        'generated_reports': GeneratedReport.objects.filter(template__kind='aan_afvoer', area = aan_afvoergebied)
+        'generated_reports': GeneratedReport.objects.filter(template__kind='aan_afvoer',
+                                                            area=aan_afvoergebied)
     })
 
     report_templates = {}
-
 
     return render_to_response(
         template,
@@ -90,7 +87,6 @@ def index(request, template='lizard_reportgenerator/index.html'):
 
         },
         context_instance=RequestContext(request))
-
 
 
 class Error(Exception):
@@ -113,13 +109,13 @@ def get_pdf_report(report_template, username, area_id=None, request=None):
     )
     report = generator(
         username,
-        format='html',
+        file_format='html',
         area_id=area_id,
         request=request,
     )
-   
+
     try:
-        encoded_report = report.encode("UTF-8") # Unicode
+        encoded_report = report.encode("UTF-8")  # Unicode
     except UnicodeEncodeError:
         encoded_report = report.encode("ISO-8859-1")  # Latin
 
@@ -134,7 +130,7 @@ def get_pdf_report(report_template, username, area_id=None, request=None):
     authenticated_opener = urllib2.build_opener()
     cookie_header = (
         'Cookie',
-        ';'.join(ck + '=' + cv 
+        ';'.join(ck + '=' + cv
                  for ck, cv in request.COOKIES.items()),
     )
     authenticated_opener.addheaders.append(cookie_header)
@@ -142,7 +138,7 @@ def get_pdf_report(report_template, username, area_id=None, request=None):
 
     # Monkeypatch httplib to send our headers:
     def monkeypatch_method(cls):
-        """ 
+        """
         By guido himself
         http://mail.python.org/pipermail/python-dev/2008-January/076194.html
         """
@@ -171,14 +167,13 @@ def get_pdf_report(report_template, username, area_id=None, request=None):
     raise Exception
 
 
-from django.utils import encoding
-
 def convert_unicode_to_string(x):
     """
     >>> convert_unicode_to_string(u'ni\xf1era')
     'niera'
     """
     return encoding.smart_str(x, encoding='ascii', errors='ignore')
+
 
 def stringify(value):
 
@@ -188,7 +183,7 @@ def stringify(value):
         else:
             return value
     except Exception, e:
-        return 'field error: %s'%e
+        return 'field error: %s' % e
 
 
 def create_csv_from_grid(grid):
@@ -198,9 +193,7 @@ def create_csv_from_grid(grid):
 
     buffer = StringIO.StringIO()
 
-
     grid = [[stringify(value) for value in row] for row in grid]
-
 
     csv_writer = csv.writer(buffer, delimiter=',',
                             quoting=csv.QUOTE_MINIMAL)
@@ -208,21 +201,21 @@ def create_csv_from_grid(grid):
     buffer.seek(0)
     return buffer
 
+
 def create_xls_from_grid(grid):
     """
         create csv from a grid with rows and cols
     """
 
-
     wb = excel.Workbook()
     ws = wb.add_sheet('rapport')
 
-    font1 = excel.Formatting.Font()
+    font1 = excel.File_Formatting.Font()
     font1.name = 'Arial'
     font1.height = 200
     font1.bold = True
 
-    font2 = excel.Formatting.Font()
+    font2 = excel.File_Formatting.Font()
     font2.name = 'Arial'
     font2.height = 160
 
@@ -245,13 +238,13 @@ def create_xls_from_grid(grid):
     for row in grid:
         c = 0
         for value in row:
-            if value != None:
+            if value is not None:
                 try:
-                    ws.write( r, c, stringify(value))
+                    ws.write(r, c, stringify(value))
                 except UnicodeEncodeError, e:
-                    ws.write( r, c, 'field error: %s'%e)
+                    ws.write(r, c, 'field error: %s' % e)
             c = c + 1
-        if c%40 == 0:
+        if c % 40 == 0:
             ws.flush_row_data()
         r = r + 1
 
@@ -262,7 +255,8 @@ def create_xls_from_grid(grid):
 
     return  buffer
 
-def generate_report(request_or_username, format='pdf', report_id=None, area_id=None, return_as_file=False):
+
+def generate_report(request_or_username, file_format='pdf', report_id=None, area_id=None, return_as_file=False):
     '''
         function calls generation functions (specified in the configuration) and returns document
     '''
@@ -276,72 +270,95 @@ def generate_report(request_or_username, format='pdf', report_id=None, area_id=N
 
     report_template = ReportTemplate.objects.get(pk=report_id)
     report_module = __import__(report_template.generation_module, fromlist='something')
+    filename = "{0}_{1}.{2}".format(report_template.slug, area_id, file_format)
     #todo, improve interaction with return, saving achieve, etc
 
-    if format == 'pdf':
+    if file_format == 'pdf':
         pdf = get_pdf_report(report_template, username, area_id, request=request)
         if return_as_file:
             return pdf
         else:
             return HttpResponse(pdf.getvalue(), mimetype='application/pdf')
 
-    elif format == 'html':
+    elif file_format == 'html':
         report = getattr(report_module, report_template.generation_function)(
             username,
-            format='html',
+            file_format='html',
             report_id=report_id,
             area_id=area_id,
             request=request
         )
         return HttpResponse(report, mimetype='text/html')
 
-    elif format == 'rtf':
-        report = getattr(report_module, report_template.generation_function)(username, format=format, report_id=report_id, area_id=area_id, request=request)
+    elif file_format == 'rtf':
+        report = getattr(report_module, report_template.generation_function)(
+            username,
+            file_format=file_format,
+            report_id=report_id,
+            area_id=area_id,
+            request=request)
         if return_as_file:
             return report
         else:
-            return HttpResponse(report, mimetype='application/rtf')
+            response = HttpResponse(report, mimetype='application/{}'.format(file_format))
+            response['Content-Disposition'] = 'attachment; filename="{}"'.format(filename)
+            return response
 
-    elif format == 'docx':
-        report = getattr(report_module, report_template.generation_function)(username, format=format, report_id=report_id, area_id=area_id, request=request)
+    elif file_format == 'docx':
+        report = getattr(report_module, report_template.generation_function)(
+            username,
+            file_format=file_format,
+            report_id=report_id,
+            area_id=area_id,
+            request=request)
         if return_as_file:
             return report
         else:
-            return HttpResponse(doc, mimetype='application/vnd.openxmlformats-officedocument.wordprocessingml.document')
+            response = HttpResponse(report, mimetype='application/vnd.openxmlfile_file_formats-officedocument.wordprocessingml.document')
+            response['Content-Disposition'] = 'attachment; filename="{}"'.format(filename)
+            return response
 
-    elif format == 'doc':
-        report = getattr(report_module, report_template.generation_function)(username, format=format, report_id=report_id, area_id=area_id, request=request)
+    elif file_format == 'doc':
+        report = getattr(report_module, report_template.generation_function)(
+            username,
+            file_format=file_format,
+            report_id=report_id,
+            area_id=area_id,
+            request=request)
         if return_as_file:
             return report
         else:
-            return HttpResponse(doc, mimetype='application/vnd.openxmlformats-officedocument.wordprocessingml.document')
+            response = HttpResponse(report, mimetype='application/vnd.openxmlfile_file_formats-officedocument.wordprocessingml.document')
+            response['Content-Disposition'] = 'attachment; filename="{}"'.format(filename)
+            return response
 
-    elif format == 'csv':
-        grid = getattr(report_module, report_template.generation_function)(username, format='grid', report_id=report_id, area_id=area_id)
+    elif file_format == 'csv':
+        grid = getattr(report_module, report_template.generation_function)(username, file_format='grid', report_id=report_id, area_id=area_id)
         csv = create_csv_from_grid(grid)
         if return_as_file:
             return csv
         else:
             response = HttpResponse(csv.read(), mimetype='application/csv')
-            response['Content-Disposition'] = 'attachment; filename="%s.csv"'%'rapport'
+            response['Content-Disposition'] = 'attachment; filename="%s.csv"' % 'rapport'
             return response
 
-    elif format == 'xls':
-        grid = getattr(report_module, report_template.generation_function)(username, format='grid', report_id=report_id, area_id=area_id)
+    elif file_format == 'xls':
+        grid = getattr(report_module, report_template.generation_function)(username, file_format='grid', report_id=report_id, area_id=area_id)
         xls = create_xls_from_grid(grid)
         if return_as_file:
             return xls
         else:
             response = HttpResponse(xls.read(), mimetype='application/xls')
-            response['Content-Disposition'] = 'attachment; filename="%s.xls"'%'rapport'
+            response['Content-Disposition'] = 'attachment; filename="%s.xls"' % 'rapport'
             return response
     else:
-        return HttpResponse('format niet ondersteund', mimetype='text/html')
+        return HttpResponse('file_file_format niet ondersteund', mimetype='text/html')
 
 
-def get_generated_report(request,generated_report_id=None,file_format='pdf'):
+def get_generated_report(request, generated_report_id=None, file_format='pdf'):
 
     generated_report = get_object_or_404(GeneratedReport, id=generated_report_id)
-    response = HttpResponse(generated_report.get_file(file_format), mimetype='application/'+file_format)
+    response = HttpResponse(generated_report.get_file(file_format),
+                            mimetype='application/' + file_format)
     response['Content-Disposition'] = 'attachment;'
     return response